+++ mex/editor/login_mex/main.py
@@ -1,9 +1,9 @@
 import reflex as rx
 
 from mex.editor.components import login_form
-from mex.editor.login_ldap.state import LoginLdapState
+from mex.editor.login_mex.state import LoginMexState
 
 
 def index() -> rx.Component:
     """Return the index for the login component."""
-    return login_form(LoginLdapState)
+    return login_form(LoginMexState)
