+++ mex/mex.py
@@ -16,8 +16,8 @@ from mex.editor.create.main import index as create_index
 from mex.editor.create.state import CreateState
 from mex.editor.edit.main import index as edit_index
 from mex.editor.edit.state import EditState
-from mex.editor.login.main import index as login_index
 from mex.editor.login_ldap.main import index as login_ldap_index
+from mex.editor.login_mex.main import index as login_mex_index
 from mex.editor.rules.state import RuleState
 from mex.editor.search.main import index as search_index
 from mex.editor.search.state import SearchState
@@ -79,7 +79,7 @@ app.add_page(
     ],
 )
 app.add_page(
-    login_index,
+    login_mex_index,
     route="/login",
     title="MEx Editor | Login",
 )
